Convert to size_t.

+++ fc-solve/source/delta_states.h
@@ -137,8 +137,8 @@ typedef struct
 #ifndef FCS_FREECELL_ONLY
     int sequences_are_built_by;
 #endif
-    int num_freecells;
-    int num_columns;
+    size_t num_freecells;
+    size_t num_columns;
     fcs_state_t *init_state, *derived_state;
     int bits_per_orig_cards_in_column;
 } fc_solve_delta_stater_t;
